Helper function to generate color mappings for Gnumber and Gnumber_2

Helper function to generate color mappings for Gnumber and Gnumber_2

Usage

generate_color_mappings(
  dt_plate_subset,
  untrt_tag = gDRutils::get_env_identifiers("untreated_tag")
)

Arguments

dt_plate_subset

The subset of data

untrt_tag

Untreated tag identifier

Value

A list containing color mappings for Gnumber and Gnumber_2

Author

Bartosz Czech czech.bartosz@external.gene.com